[Gold II] Title: 카드 뒤집기 게임, Time: 196 ms, Memory: 325804 KB -BaekjoonHub

+++ 백준/Gold/22026. 카드 뒤집기 게임/카드 뒤집기 게임.cc
@@ -1,7 +1,7 @@
 #include <bits/stdc++.h>
 
 using namespace std;
-typedef long long ll;
+typedef int ll;
 
 ll num[2010][2010];
 vector<ll> vec[5000010], yuk[5000010];
